requirements.txt
streamlit>=1.29.0
mysql-connector-python>=8.1.0
reportlab>=4.0.0
python-dotenv>=1.0.0
bcrypt>=4.1.2